Version 1.6
Added one new feature: When you use :BlogSend, your default browser will open either the preview URL or the actual URL in your default browser (I have not tested this in Windows). If it is a published post, it will use the "?p=id" URL format. Note that this doesn't affect how you have your wordpress default URL slug setup.

You can configure the plugin to not open a browser window by changing the enable_preview setting to 0. Currently, it will open a new window/tab everytime you do :BlogSend. I hope to modify this to only do it once per post in the future.

I also now show the appropriate url in the Preview line based on whether it is a draft or published post.